Move plugin check to WPComPushNotificationsBenefitsViewModel
