+++ WooCommerce/Classes/ViewRelated/Dashboard/DashboardViewHostingController.swift
@@ -372,8 +372,6 @@ private extension DashboardViewHostingController {
         rootView.onConnectWPComSetup = { [weak self] in
             guard let self else { return }
             let benefitsViewModel = WPComPushNotificationsBenefitsViewModel(
-                variant: self.viewModel.isWooPluginOutdated ? .pluginUpdate : .connect,
-                pluginVersion: self.viewModel.outdatedPluginVersion,
                 onDismiss: {
                     self.dismiss(animated: true)
                 }
