+++ WooCommerce/Classes/ViewRelated/Dashboard/DashboardViewModel.swift
@@ -77,9 +77,6 @@ final class DashboardViewModel: ObservableObject {
 
     @Published private(set) var shouldSuggestWPComConnection = false
 
-    @Published private(set) var isWooPluginOutdated = false
-    private(set) var outdatedPluginVersion: String = ""
-
     @Published private(set) var dismissedWPComConnectionSuggestion = false
 
     @Published private var hasOrders = false
@@ -107,7 +104,6 @@ final class DashboardViewModel: ObservableObject {
     private let usageTracksEventEmitter: StoreStatsUsageTracksEventEmitter
     private let blazeLocalNotificationScheduler: BlazeLocalNotificationScheduler
     private let tapToPayAwarenessMomentDeterminer: TapToPayAwarenessMomentDetermining
-    private let pluginVersionChecker: PluginVersionCheckerProtocol?
     private let clientSideBannerProvider: ClientSideBannerProvider
 
     private var subscriptions: Set<AnyCancellable> = []
@@ -162,7 +158,6 @@ final class DashboardViewModel: ObservableObject {
          siteIsCIABEligibilityChecker: CIABEligibilityCheckerProtocol = CIABEligibilityChecker(),
          localNotificationScheduler: BlazeLocalNotificationScheduler? = nil,
          tapToPayAwarenessMomentDeterminer: TapToPayAwarenessMomentDetermining = TapToPayAwarenessMomentDeterminer(),
-         pluginVersionChecker: PluginVersionCheckerProtocol? = nil,
          clientSideBannerProvider: ClientSideBannerProvider? = nil) {
         self.siteID = siteID
         self.stores = stores
@@ -206,16 +201,6 @@ final class DashboardViewModel: ObservableObject {
         self.blazeLocalNotificationScheduler.observeNotificationUserResponse()
 
         self.tapToPayAwarenessMomentDeterminer = tapToPayAwarenessMomentDeterminer
-        self.pluginVersionChecker = pluginVersionChecker ?? {
-            guard let site = stores.sessionManager.defaultSite, site.isJetpackConnected else {
-                return nil
-            }
-            return PluginVersionChecker(
-                siteID: site.siteID,
-                pluginPath: WooPluginRequirements.pluginPath,
-                minimumVersion: WooPluginRequirements.minimumVersion
-            )
-        }()
 
         self.clientSideBannerProvider = clientSideBannerProvider ?? ClientSideBannerProvider(
             stores: stores,
@@ -425,7 +410,6 @@ private extension DashboardViewModel {
             }
             group.addTask { [weak self] in
                 await self?.updateSelfDrivenPushRegistrationStatus()
-                await self?.checkWooPluginVersion()
             }
         }
     }
@@ -534,11 +518,11 @@ private extension DashboardViewModel {
             })
             .store(in: &subscriptions)
 
-        $dashboardCards.combineLatest($isInAppFeedbackCardVisible, $shouldSuggestWPComConnection, $isWooPluginOutdated)
+        $dashboardCards.combineLatest($isInAppFeedbackCardVisible, $shouldSuggestWPComConnection)
             .combineLatest($showNewCardsNotice, $hasOrders, $isReloadingAllData)
             .sink { [weak self] combinedResult in
                 guard let self else { return }
-                let ((cards, showFeedbackCard, suggestWPComConnection, isWooPluginOutdated), showNewCardsNotice, hasOrders, isReloading) = combinedResult
+                let ((cards, showFeedbackCard, suggestWPComConnection), showNewCardsNotice, hasOrders, isReloading) = combinedResult
                 let cardsToShow: [DashboardCard] = {
                     var allCards = cards.filter { $0.availability == .show && $0.enabled }
 
@@ -558,8 +542,8 @@ private extension DashboardViewModel {
                         allCards.append(DashboardCard.shareStoreCard)
                     }
 
-                    /// Insert card for connecting WPCom or updating plugin at the top if needed
-                    if suggestWPComConnection || isWooPluginOutdated {
+                    /// Insert card for connecting WPCom at the top if needed
+                    if suggestWPComConnection {
                         allCards.insert(DashboardCard.connectWPCom, at: 0)
                     }
                     return allCards
@@ -957,21 +941,6 @@ private extension DashboardViewModel {
             featureFlagService.isFeatureFlagEnabled(.selfDrivenPushTokenAppPasswords)
     }
 
-    @MainActor
-    func checkWooPluginVersion() async {
-        isWooPluginOutdated = false
-        outdatedPluginVersion = ""
-        guard let pluginVersionChecker else { return }
-        do {
-            let result = try await pluginVersionChecker.checkCompatibility()
-            if case .incompatible(let currentVersion, _) = result {
-                isWooPluginOutdated = true
-                outdatedPluginVersion = currentVersion
-            }
-        } catch {
-            DDLogError("⛔️ Plugin version check failed: \(error)")
-        }
-    }
 }
 
 // MARK: InAppFeedback card
